+++ flink-table/flink-table-runtime-blink/src/main/java/org/apache/flink/table/dataformat/DataFormatConverters.java
@@ -33,6 +33,7 @@ import org.apache.flink.api.java.typeutils.RowTypeInfo;
 import org.apache.flink.api.java.typeutils.TupleTypeInfo;
 import org.apache.flink.api.java.typeutils.TupleTypeInfoBase;
 import org.apache.flink.api.java.typeutils.runtime.TupleSerializerBase;
+import org.apache.flink.table.runtime.functions.DateTimeUtils;
 import org.apache.flink.table.type.InternalType;
 import org.apache.flink.table.type.TypeConverters;
 import org.apache.flink.table.typeutils.BaseRowTypeInfo;
@@ -48,7 +49,6 @@ import java.sql.Timestamp;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.TimeZone;
 
 import scala.Product;
 
@@ -387,65 +387,6 @@ public class DataFormatConverters {
 		}
 	}
 
-	public static final TimeZone LOCAL_TZ = TimeZone.getDefault();
-	public static final long MILLIS_PER_DAY = 86400000L; // = 24 * 60 * 60 * 1000
-
-	/** Converts the Java type used for UDF parameters of SQL TIME type
-	 * ({@link java.sql.Time}) to internal representation (int).
-	 *
-	 * <p>Converse of {@link #internalToTime(int)}. */
-	public static int toInt(java.sql.Time v) {
-		return (int) (toLong(v) % MILLIS_PER_DAY);
-	}
-
-	/** Converts the Java type used for UDF parameters of SQL DATE type
-	 * ({@link java.sql.Date}) to internal representation (int).
-	 *
-	 * <p>Converse of {@link #internalToDate(int)}. */
-	public static int toInt(java.util.Date v) {
-		return toInt(v, LOCAL_TZ);
-	}
-
-	public static int toInt(java.util.Date v, TimeZone timeZone) {
-		return (int) (toLong(v, timeZone) / MILLIS_PER_DAY);
-	}
-
-	public static long toLong(java.util.Date v) {
-		return toLong(v, LOCAL_TZ);
-	}
-
-	/** Converts the Java type used for UDF parameters of SQL TIMESTAMP type
-	 * ({@link java.sql.Timestamp}) to internal representation (long).
-	 *
-	 * <p>Converse of {@link #internalToTimestamp(long)}. */
-	public static long toLong(Timestamp v) {
-		return toLong(v, LOCAL_TZ);
-	}
-
-	public static long toLong(java.util.Date v, TimeZone timeZone) {
-		long time = v.getTime();
-		return time + (long) timeZone.getOffset(time);
-	}
-
-	/** Converts the internal representation of a SQL DATE (int) to the Java
-	 * type used for UDF parameters ({@link java.sql.Date}). */
-	public static java.sql.Date internalToDate(int v) {
-		final long t = v * MILLIS_PER_DAY;
-		return new java.sql.Date(t - LOCAL_TZ.getOffset(t));
-	}
-
-	/** Converts the internal representation of a SQL TIME (int) to the Java
-	 * type used for UDF parameters ({@link java.sql.Time}). */
-	public static java.sql.Time internalToTime(int v) {
-		return new java.sql.Time(v - LOCAL_TZ.getOffset(v));
-	}
-
-	/** Converts the internal representation of a SQL TIMESTAMP (long) to the Java
-	 * type used for UDF parameters ({@link java.sql.Timestamp}). */
-	public static java.sql.Timestamp internalToTimestamp(long v) {
-		return new java.sql.Timestamp(v - LOCAL_TZ.getOffset(v));
-	}
-
 	/**
 	 * Converter for date.
 	 */
@@ -457,12 +398,12 @@ public class DataFormatConverters {
 
 		@Override
 		Integer toInternalImpl(Date value) {
-			return toInt(value);
+			return DateTimeUtils.toInt(value);
 		}
 
 		@Override
 		Date toExternalImpl(Integer value) {
-			return internalToDate(value);
+			return DateTimeUtils.internalToDate(value);
 		}
 
 		@Override
@@ -482,12 +423,12 @@ public class DataFormatConverters {
 
 		@Override
 		Integer toInternalImpl(Time value) {
-			return toInt(value);
+			return DateTimeUtils.toInt(value);
 		}
 
 		@Override
 		Time toExternalImpl(Integer value) {
-			return internalToTime(value);
+			return DateTimeUtils.internalToTime(value);
 		}
 
 		@Override
@@ -507,12 +448,12 @@ public class DataFormatConverters {
 
 		@Override
 		Long toInternalImpl(Timestamp value) {
-			return toLong(value);
+			return DateTimeUtils.toLong(value);
 		}
 
 		@Override
 		Timestamp toExternalImpl(Long value) {
-			return internalToTimestamp(value);
+			return DateTimeUtils.internalToTimestamp(value);
 		}
 
 		@Override
